feat: implement reasoning actions with slot-fill, available_when, and set write-backs
- Generate per-reasoning-action named async wrappers instead of
  functools-based _make_tool closure for correct LLM tool schemas
- Slot-fill: only expose params with value '...' to the LLM; pre-bound
  params are resolved from state or inlined as literals
- Support available_when guards that skip tool execution when condition
  is false
- Support set write-backs that store action outputs into state variables
  after execution
- Support transition targets (@utils.transition to @subagent.X)
- Add _set_variables tool so LLM can set mutable variables by description
- Add _pending_transition mechanism for inter-subagent transitions
- Fix type mapping: number -> float (IEEE 754), add integer, long,
  currency, date, datetime, time, timestamp, id types
- Add flow tests for case_escalation_bot scenario

+++ tests/generator/tool-generator.test.ts
@@ -55,7 +55,7 @@ describe('ToolGenerator', () => {
   it('maps type hints', () => {
     const gen = new ToolGenerator();
     expect(gen.mapParamType('string')).toBe('str');
-    expect(gen.mapParamType('number')).toBe('int');
+    expect(gen.mapParamType('number')).toBe('float');
     expect(gen.mapParamType('boolean')).toBe('bool');
     expect(gen.mapParamType('object')).toBe('dict');
   });
